requirements.txt
click>=8.2.1
google-genai>=0.3.0
google-api-core>=2.11.0
anyio>=4.8.0,<5.0.0
tenacity>=8.2.3
python-dotenv>=1.0.0
pandas>=2.1.0,<3.0.0
pdfplumber>=0.10.0
numpy>=2.0.0,<2.3.0
faiss-cpu>=1.7.4
setuptools>=68.0.0
